+++ src/Javinizer/Private/Get-MetadataNfo.ps1
@@ -10,6 +10,8 @@ function Get-MetadataNfo {
     begin {
         $javlibraryGenres = @()
         $r18Genres = @()
+        $actressObject = @()
+        $r18CsvPath = Join-Path -Path $ScriptRoot -ChildPath 'r18-thumbs.csv'
 
         if ($Settings.Metadata.'normalize-genres' -eq 'True') {
             try {
@@ -84,6 +86,36 @@ function Get-MetadataNfo {
 
         if ($DataObject.Actress.Count -gt 0) {
             if ($DataObject.Actress.Count -eq 1) {
+                if (-not ($R18ThumbCsv.FullName -like $DataObject.Actress)) {
+                    if (($DataObject.ActressThumbUrl -notlike '*nowprinting*') -or ($null -ne $DataObject.ActressThumbUrl)) {
+                        $actressFirstName, $actressLastName = $DataObject.Actress -split ' '
+                        $actressFullName = $actressFirstName + ' ' + $actressLastName
+                        $actressFullNameReversed = $actressLastName + ' ' + $actressFirstName
+                        $actressThumbUrl = ($DataObject.ActressThumbUrl).ToString()
+                        $actressObject += [pscustomobject]@{
+                            FirstName        = $actressFirstName.Trim()
+                            LastName         = $actressLastName.Trim()
+                            FullName         = $actressFullName.Trim()
+                            FullNameReversed = $actressFullNameReversed.Trim()
+                            ThumbUrl         = (@($DataObject.ActressThumbUrl) | Out-String).Trim()
+                            Alias            = ''
+                        }
+
+                        try {
+                            $actressObject | Export-Csv -LiteralPath $r18CsvPath -Append -NoTypeInformation
+                        } catch {
+                            Write-Warning "[$($MyInvocation.MyCommand.Name)] Error appending actress to [$r18CsvPath], waiting 2 seconds and trying again"
+                            Start-Sleep -Seconds 2
+                            try {
+                                $actressObject | Export-Csv -LiteralPath $r18CsvPath -Append
+                            } catch {
+                                Write-Warning "[$($MyInvocation.MyCommand.Name)] Error appending actress to [$r18CsvPath], skipping"
+                            }
+                        }
+                        Write-Verbose "[$($MyInvocation.MyCommand.Name)] Actress [$($DataObject.Actress)] written to [$r18CsvPath]"
+                    }
+                }
+
                 if (($DataObject.ActressThumbUrl -like '*nowprinting*') -or ($null -eq $DataObject.ActressThumbUrl)) {
                     if (($csvFullName -like $DataObject.Actress) -or ($csvFullNameAlias -like $DataObject.Actress)) {
                         $index = $csvFullname.IndexOf("$($DataObject.Actress)")
@@ -112,6 +144,26 @@ function Get-MetadataNfo {
 "@
             } else {
                 for ($i = 0; $i -lt $DataObject.Actress.Count; $i++) {
+                    if (-not ($R18ThumbCsv.FullName -like $DataObject.Actress[$i])) {
+                        if (($DataObject.ActressThumbUrl[$i] -notlike '*nowprinting*') -or ($null -ne $DataObject.ActressThumbUrl[$i])) {
+                            $actressFirstName, $actressLastName = $DataObject.Actress[$i] -split ' '
+                            $actressFullName = $actressFirstName + ' ' + $actressLastName
+                            $actressFullNameReversed = $actressLastName + ' ' + $actressFirstName
+
+                            $actressObject = [pscustomobject]@{
+                                FirstName        = $actressFirstName.Trim()
+                                LastName         = $actressLastName.Trim()
+                                FullName         = $actressFullName.Trim()
+                                FullNameReversed = $actressFullNameReversed.Trim()
+                                ThumbUrl         = $DataObject.ActressThumbUrl[$i]
+                                Alias            = ''
+                            }
+
+                            $actressObject | Export-Csv -LiteralPath $r18CsvPath -Append -NoTypeInformation
+                            Write-Verbose "[$($MyInvocation.MyCommand.Name)] Actress [$($DataObject.Actress[$i])] written to [$r18CsvPath]"
+                        }
+                    }
+
                     if ($null -eq $DataObject.ActressThumbUrl) {
                         # Create empty array amounting to number of actresses found if scraped from javlibrary
                         # This will allow matching actresses from r18 thumb csv
@@ -162,3 +214,16 @@ function Get-MetadataNfo {
         Write-Output $nfoString
     }
 }
+
+
+<#
+
+if match actress + ActressThumbUrl
+check r18 csv
+    if not in r18 csv
+        add row
+    else
+        don't do anything
+
+
+#>